move account_manager to beta
